requirements.txt
langchain==0.3.0
langchain-community==0.3.0
fastapi==0.111.1
uvicorn==0.30.3
python-dotenv
textblob
huggingface-hub